backend/requirements.txt
django == 4.2
djangorestframework == 3.15.2
djangorestframework-simplejwt == 5.4.0
django-cors-headers == 4.7.0
faker == 37.0.0
django_extensions == 4.1